% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{list_dimred_methods}
\alias{list_dimred_methods}
\alias{dimred}
\alias{dimred_pca}
\alias{dimred_mds}
\alias{dimred_mds_sammon}
\alias{dimred_mds_isomds}
\alias{dimred_mds_smacof}
\alias{dimred_dm_diffusionMap}
\alias{dimred_ica}
\alias{dimred_lle}
\title{Perform simple dimensionality reduction}
\usage{
list_dimred_methods()

dimred(x, method, ndim, ...)

dimred_pca(x, ndim = 2)

dimred_mds(x, ndim = 2)

dimred_mds_sammon(x, ndim = 2)

dimred_mds_isomds(x, ndim = 2)

dimred_mds_smacof(x, ndim = 2)

dimred_dm_diffusionMap(x, ndim = 2)

dimred_ica(x, ndim = 3)

dimred_lle(x, ndim = 3)
}
\arguments{
\item{x}{Log transformed expression data, with rows as cells and columns as features}

\item{method}{The name of the dimensionality reduction method to use}

\item{ndim}{The number of dimensions}

\item{...}{Any arguments to be passed to the dimensionality reduction method}
}
\description{
Perform simple dimensionality reduction
}
