% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{dimred_mds_landmark}
\alias{dimred_mds_landmark}
\title{Landmark MDS}
\usage{
dimred_mds_landmark(x, ndim = 2, landmark_method = "naive",
  num_landmarks = 100, rescale = T)
}
\arguments{
\item{x}{a numeric matrix}

\item{ndim}{the maximum dimension of the space which the data are to be represented in; must be in {1, 2, \ldots, n-1}.}

\item{landmark_method}{Must be "naive" for now. Other landmark methods will be supported in the future.}

\item{num_landmarks}{the number of landmarks to be selected.}

\item{rescale}{A logical indicating whether or not the returned space should be rescaled and centered.}
}
\description{
Landmark MDS
}
\seealso{
[SCORPIUS::reduce_dimensionality()]
}
