% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{dimred_tsne}
\alias{dimred_tsne}
\title{tSNE}
\usage{
dimred_tsne(x, ndim = 2, perplexity = 30, theta = 0.5,
  initial_dims = 50)
}
\arguments{
\item{perplexity}{numeric; Perplexity parameter}

\item{theta}{numeric; Speed/accuracy trade-off (increase for less accuracy), set to 0.0 for exact TSNE (default: 0.5)}

\item{initial_dims}{integer; the number of dimensions that should be retained in the initial PCA step (default: 50)}
}
\description{
tSNE
}
\seealso{
\code{\link[Rtsne:Rtsne]{Rtsne::Rtsne()}}
}
