% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{dyndimred}
\alias{dyndimred}
\title{Common dimensionality reduction methods}
\description{
Provides a common interface for applying common dimensionality reduction methods,
Such as PCA, ICA, diffusion maps, LLE, t-SNE, and umap.
}
