% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{dimred}
\alias{dimred}
\alias{list_dimred_methods}
\alias{dimred_pca}
\alias{dimred_mds}
\alias{dimred_mds_sammon}
\alias{dimred_mds_isomds}
\alias{dimred_mds_smacof}
\alias{dimred_tsne}
\alias{dimred_dm_diffusionMap}
\alias{dimred_ica}
\alias{dimred_lle}
\alias{dimred_umap}
\title{Perform simple dimensionality reduction}
\usage{
dimred(x, method, ndim, ...)

list_dimred_methods()

dimred_pca(x, ndim = 3)

dimred_mds(x, ndim = 3)

dimred_mds_sammon(x, ndim = 3)

dimred_mds_isomds(x, ndim = 3)

dimred_mds_smacof(x, ndim = 3)

dimred_tsne(x, ndim = 3)

dimred_dm_diffusionMap(x, ndim = 3)

dimred_ica(x, ndim = 3)

dimred_lle(x, ndim = 3)

dimred_umap(x, ndim = 2, n_neighbors = 15L)
}
\arguments{
\item{x}{Log transformed expression data, with rows as cells and columns as features}

\item{method}{The name of the dimensionality reduction method to use}

\item{ndim}{The number of dimensions}

\item{...}{Any arguments to be passed to the dimensionality reduction method}

\item{n_neighbors}{integer. The size of local neighborhood 
(in terms of number of neighboring sample points) used for manifold 
approximation. Larger values result in more global views of the manifold, 
while smaller values result in more local data being preserved. In general 
values should be in the range 2 to 100.}
}
\description{
Perform simple dimensionality reduction
}
