% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{list_dimred_methods}
\alias{list_dimred_methods}
\alias{dimred}
\alias{dimred_pca}
\alias{dimred_mds}
\alias{dimred_mds_sammon}
\alias{dimred_mds_isomds}
\alias{dimred_mds_smacof}
\alias{dimred_tsne}
\alias{dimred_dm_diffusionMap}
\alias{dimred_ica}
\alias{dimred_lle}
\alias{dimred_umap}
\title{Perform simple dimensionality reduction}
\usage{
list_dimred_methods()

dimred(x, method, ndim, ...)

dimred_pca(x, ndim = 2)

dimred_mds(x, ndim = 2)

dimred_mds_sammon(x, ndim = 2)

dimred_mds_isomds(x, ndim = 2)

dimred_mds_smacof(x, ndim = 2)

dimred_tsne(x, ndim = 2)

dimred_dm_diffusionMap(x, ndim = 2)

dimred_ica(x, ndim = 2)

dimred_lle(x, ndim = 2)

dimred_umap(x, ndim = 2, n_neighbors = 15L, alpha = 1,
  init = "spectral")
}
\arguments{
\item{x}{Log transformed expression data, with rows as cells and columns as features}

\item{method}{The name of the dimensionality reduction method to use}

\item{ndim}{The number of dimensions}

\item{...}{Any arguments to be passed to the dimensionality reduction method}

\item{n_neighbors}{The size of local neighborhood (in terms of number of
neighboring sample points) used for manifold approximation. Larger values
result in more global views of the manifold, while smaller values result in
more local data being preserved. In general values should be in the range
\code{2} to \code{100}.}

\item{alpha}{Initial learning rate used in optimization of the coordinates.}

\item{init}{Type of initialization for the coordinates. Options are:
\itemize{
  \item \code{"spectral"} Spectral embedding using the normalized Laplacian
  of the fuzzy 1-skeleton, with Gaussian noise added.
  \item \code{"normlaplacian"}. Spectral embedding using the normalized
  Laplacian of the fuzzy 1-skeleton, without noise.
  \item \code{"random"}. Coordinates assigned using a uniform random
  distribution between -10 and 10.
  \item \code{"lvrandom"}. Coordinates assigned using a Gaussian
  distribution with standard deviation 1e-4, as in LargeVis
  (Tang et al., 2016).
  \item \code{"laplacian"}. Spectral embedding using the Laplacian Eigenmap
  (Belkin and Niyogi, 2002).
  \item \code{"spca"}. The first two principal components from PCA of
  \code{X} if \code{X} is a data frame, and from a 2-dimensional classical
  MDS if \code{X} is of class \code{"dist"}. These vectors are then scaled
  so their standard deviation is 0.0001.
  \item A matrix of initial coordinates.
}}
}
\description{
Perform simple dimensionality reduction
}
