% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred.R, R/dimred_dm_destiny.R,
%   R/dimred_dm_diffusionmap.R, R/dimred_ica.R, R/dimred_lle.R, R/dimred_mds.R,
%   R/dimred_mds_isomds.R, R/dimred_mds_sammon.R, R/dimred_mds_smacof.R,
%   R/dimred_pca.R, R/list_dimred_methods.R
\name{dimred}
\alias{dimred}
\alias{dimred_dm_destiny}
\alias{dimred_dm_diffusionmap}
\alias{dimred_ica}
\alias{dimred_lle}
\alias{dimred_mds}
\alias{dimred_mds_isomds}
\alias{dimred_mds_sammon}
\alias{dimred_mds_smacof}
\alias{dimred_pca}
\alias{list_dimred_methods}
\title{Perform simple dimensionality reduction}
\usage{
dimred(x, method, ndim, ...)

dimred_dm_destiny(x, ndim = 2)

dimred_dm_diffusionmap(x, ndim = 2)

dimred_ica(x, ndim = 3)

dimred_lle(x, ndim = 3)

dimred_mds(x, ndim = 2)

dimred_mds_isomds(x, ndim = 2)

dimred_mds_sammon(x, ndim = 2)

dimred_mds_smacof(x, ndim = 2)

dimred_pca(x, ndim = 2)

list_dimred_methods()
}
\arguments{
\item{x}{Log transformed expression data, with rows as cells and columns as features}

\item{method}{The name of the dimensionality reduction method to use}

\item{ndim}{The number of dimensions}

\item{...}{Any arguments to be passed to the dimensionality reduction method}
}
\description{
Perform simple dimensionality reduction
}
