% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{dimred_landmark_mds}
\alias{dimred_landmark_mds}
\title{Landmark MDS}
\usage{
dimred_landmark_mds(x, ndim = 2, landmark_method = "naive",
  num_landmarks = 100, rescale = T)
}
\arguments{
\item{x}{Log transformed expression data, with rows as cells and columns as features}

\item{ndim}{The number of dimensions}

\item{landmark_method}{Must be "naive" for now. Other landmark methods will be supported in the future.}

\item{num_landmarks}{the number of landmarks to be selected.}

\item{rescale}{A logical indicating whether or not the returned space should be rescaled and centered.}
}
\description{
Landmark MDS
}
\seealso{
\code{\link[SCORPIUS:reduce_dimensionality]{SCORPIUS::reduce_dimensionality()}}
}
