% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred_landmark_mds.R
\name{dimred_landmark_mds}
\alias{dimred_landmark_mds}
\title{Landmark MDS}
\usage{
dimred_landmark_mds(x, ndim = 2, distance_metric = c("spearman",
  "euclidean", "manhattan", "angular", "pearson", "kendall"),
  landmark_method = c("naive"), num_landmarks = 100)
}
\arguments{
\item{x}{Log transformed expression data, with rows as cells and columns as features}

\item{ndim}{The number of dimensions}
}
\description{
Landmark MDS
}
