% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred_landmark_mds.R
\name{dimred_landmark_mds}
\alias{dimred_landmark_mds}
\title{Landmark MDS}
\usage{
dimred_landmark_mds(x, ndim = 2, distance_metric = c("spearman",
  "euclidean", "manhattan", "angular", "pearson", "kendall"),
  num_landmarks = 500)
}
\arguments{
\item{x}{Log transformed expression data, with rows as cells and columns as features}

\item{ndim}{The number of dimensions}

\item{distance_metric}{Which distance function to use. Must be one of: \code{"spearman"}, \code{"pearson"}, \code{"kendall"}, \code{"angular"}, \code{"euclidean"} or \code{"manhattan"}.}

\item{num_landmarks}{The number of landmarks to use,}
}
\description{
Landmark MDS
}
